requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.11.3
amqp==5.1.1
annotated-types==0.5.0
anyio==3.6.2
APScheduler==3.10.4
asgiref==3.7.2
astroid==3.0.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
autopep8==2.0.2
bcrypt==4.0.1
bidict==0.22.1
billiard==3.6.4.0
bitarray==2.9.2
blinker==1.6.2
blis==0.7.11
boto3==1.28.80
botocore==1.31.80
Brotli==1.1.0
cachetools==5.3.3
catalogue==2.0.10
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptography==40.0.2
cymem==2.0.8
cytoolz==0.12.3
dill==0.3.7
distlib==0.3.8
dnspython==2.3.0
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.1
filelock==3.13.3
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hexbytes==0.3.1
hpack==4.0.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
hyperframe==6.0.1
identify==2.5.36
idna==3.4
isort==5.12.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.2.4
langcodes==3.3.0
lru-dict==1.2.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
mccabe==0.7.0
motor==3.1.2
multidict==6.0.4
murmurhash==1.0.10
nodeenv==1.9.1
numpy==1.26.4
orjson==3.9.9
packaging==24.1
parsimonious==0.9.0
passlib==1.7.4
platformdirs==3.11.0
pluggy==1.4.0
pre-commit==3.7.1
preshed==3.0.9
prompt-toolkit==3.0.38
protobuf==5.26.0
psycopg2-binary==2.9.6
pycares==4.4.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
pydash==7.0.6
PyJWT==2.6.0
pylint==3.0.2
pymongo==4.3.3
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-multipart==0.0.6
python-socketio==5.8.0
pytz==2023.3
pyunormalize==15.1.0
PyYAML==5.3.1
qrcode==7.4.2
redis==5.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.18.0
s3transfer==0.7.0
schedule==1.2.0
sentry-sdk==1.40.6
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.12
srsly==2.4.8
starlette==0.27.0
thinc==8.2.3
threadpoolctl==3.4.0
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.1
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.7.1
tzlocal==5.0.1
ua-parser==0.18.0
ujson==5.9.0
Unidecode==1.3.7
urllib3==1.26.18
user-agents==2.2.0
uvicorn==0.21.1
uvloop==0.17.0
vine==5.0.0
virtualenv==20.25.1
wasabi==1.1.2
watchfiles==0.19.0
wcwidth==0.2.6
weasel==0.3.4
web3==6.15.1
websockets==11.0.1
yarl==1.9.4
zope.event==5.0
zope.interface==6.2
